refactor(nomic): extract dev coordination verification helpers
